refactor(cli): move iOS commands to their own module
